设备可视化化部分：后台获取单个光盘库信息测试通过

+++ src/main/java/com/shdd/cfs/dto/device/tape/TapeLibraryStorageSystemStoresDetail.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class TapeLibraryStorageSystemStoresDetail {
     private String name;
-    private Integer  capacity;
-    private Integer used;
-    private Integer free;
+    private Double  capacity;
+    private Double used;
+    private Double free;
     private Integer status;
 }
